add opensearch store and hybrid store

+++ lazyllm/tools/rag/store/vector/milvus_store.py
@@ -5,14 +5,13 @@ from packaging import version
 from urllib import parse
 from typing import Dict, List, Union, Optional, Set
 
-from lazyllm import LOG
+from lazyllm import LOG, once_wrapper
 from lazyllm.thirdparty import pymilvus
 from lazyllm.common import override
 
-from ..store_base import (LazyLLMStoreBase, BUILDIN_GLOBAL_META_DESC, StoreCapability,
-                          GLOBAL_META_KEY_PREFIX, EMBED_PREFIX)
-from ..data_type import DataType
-from ..global_metadata import GlobalMetadataDesc
+from ..store_base import LazyLLMStoreBase, StoreCapability, GLOBAL_META_KEY_PREFIX, EMBED_PREFIX
+from ...data_type import DataType
+from ...global_metadata import GlobalMetadataDesc
 
 MILVUS_UPSERT_BATCH_SIZE = 500
 MILVUS_PAGINATION_OFFSET = 1000
@@ -30,25 +29,22 @@ BUILTIN_KEYS = {
 
 class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
     def __init__(self, uri: str = "", db_name: str = 'lazyllm', index_kwargs: Optional[Union[Dict, List]] = None,
-                 client_kwargs: Optional[Dict] = {}, embed_dims: Optional[Dict[str, int]] = {},
-                 embed_datatypes: Optional[Dict[str, DataType]] = {},
-                 global_metadata_desc: Optional[Dict[str, GlobalMetadataDesc]] = None):
+                 client_kwargs: Optional[Dict] = {}, ):
         # one database, different collection for each group (for standalone, add prefix to collection name)
         # when there's data need upsert, collection creation happen.
         self._uri = uri
         self._db_name = db_name
         self._index_kwargs = index_kwargs
         self._client_kwargs = client_kwargs
-        self._embed_dims = embed_dims
-        self._embed_datatypes = embed_datatypes
-        if global_metadata_desc:
-            self._global_metadata_desc = global_metadata_desc | BUILDIN_GLOBAL_META_DESC
-        else:
-            self._global_metadata_desc = BUILDIN_GLOBAL_META_DESC
         self._primary_key = 'uid'
 
     @override
-    def lazy_init(self):
+    @once_wrapper(reset_on_pickle=True)
+    def lazy_init(self, embed_dims: Optional[Dict[str, int]] = {}, embed_datatypes: Optional[Dict[str, DataType]] = {},
+                  global_metadata_desc: Optional[Dict[str, GlobalMetadataDesc]] = None, **kwargs):
+        self._embed_dims = embed_dims
+        self._embed_datatypes = embed_datatypes
+        self._global_metadata_desc = global_metadata_desc
         self._client = pymilvus.MilvusClient(uri=self._uri, **self._client_kwargs)
         if self._uri and parse.urlparse(self._uri).scheme.lower() not in ["unix", "http", "https", "tcp", "grpc"]:
             self._type = 'local'
@@ -97,7 +93,13 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
         """ delete data from the store """
         try:
             self._check_connection()
-            self._client.delete(collection_name=collection_name, **self._construct_criteria(criteria))
+            if not self._client.has_collection(collection_name):
+                return True
+            self._client.load_collection(collection_name)
+            if len(criteria) == 0:
+                self._client.drop_collection(collection_name=collection_name)
+            else:
+                self._client.delete(collection_name=collection_name, **self._construct_criteria(criteria))
             return True
         except Exception as e:
             LOG.error(f'[Milvus Store - delete] error: {e}')
@@ -108,6 +110,9 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
         """ get data from the store """
         try:
             self._check_connection()
+            if not self._client.has_collection(collection_name):
+                return []
+            self._client.load_collection(collection_name)
             col_desc = self._client.describe_collection(collection_name=collection_name)
             field_names = [field.get("name") for field in col_desc.get('fields', [])
                            if field.get("name").startswith(EMBED_PREFIX)]
@@ -147,16 +152,18 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
         for k, kws in BUILTIN_KEYS.items():
             field_list.append(pymilvus.FieldSchema(name=k, **kws))
         for k, desc in self._global_metadata_desc.items():
+            field_name = self._gen_global_meta_key(k)
             if desc.data_type == DataType.ARRAY:
-                if not desc.element_type:
-                    raise ValueError(f'Milvus field [{k}]: `element_type` is required when `data_type` is ARRAY.')
+                if desc.element_type is None:
+                    raise ValueError(f'Milvus field [{field_name}]: '
+                                     '`element_type` is required when `data_type` is ARRAY.')
                 field_args = {'element_type': TYPE2MILVUS[desc.element_type], 'max_capacity': desc.max_size}
                 if desc.element_type == DataType.VARCHAR: field_args['max_length'] = 65535
             elif desc.data_type == DataType.VARCHAR:
                 field_args = {'max_length': desc.max_size}
             else:
                 field_args = {}
-            field_list.append(pymilvus.FieldSchema(name=k, dtype=TYPE2MILVUS[desc.data_type],
+            field_list.append(pymilvus.FieldSchema(name=field_name, dtype=TYPE2MILVUS[desc.data_type],
                                                    default_value=desc.default_value, **field_args))
         return field_list
 
@@ -176,7 +183,7 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
                     if embed_key == k:
                         index_kwarg = item.copy()
                         index_kwarg.pop('embed_key', None)
-                        index_params.add_index(field_name=k, **index_kwarg)
+                        index_params.add_index(field_name=embed_field_name, **index_kwarg)
                         break
             elif isinstance(self._index_kwargs, dict):
                 index_params.add_index(field_name=k, **self._index_kwargs)
@@ -201,7 +208,8 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
     def _deserialize_data(self, d: dict) -> dict:
         """ deserialize data from vector store """
         res = {
-            self._primary_key: d.get(self._primary_key, '')
+            self._primary_key: d.get(self._primary_key, ''),
+            "embedding": {}
         }
         for k, v in d.items():
             if k.startswith(EMBED_PREFIX):
@@ -217,6 +225,7 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
     def _construct_criteria(self, criteria: dict) -> dict:
         """ construct criteria for delete """
         res = {}
+        criteria = dict(criteria)
         if self._primary_key in criteria:
             res["ids"] = criteria[self._primary_key]
         else:
@@ -256,6 +265,8 @@ class MilvusStore(LazyLLMStoreBase, capability=StoreCapability.VECTOR):
 
     def _construct_filter_expr(self, filters: Dict[str, Union[str, int, List, Set]]) -> str:
         ret_str = ""
+        if not filters:
+            return ret_str
         for name, candidates in filters.items():
             desc = self._global_metadata_desc.get(name)
             if not desc:
